Move compare source to utils and export it
